update to 0.21.1

+++ PKGBUILD
@@ -2,16 +2,16 @@
 
 pkgname=mingw-w64-libpsl
 _pkgname=libpsl
-pkgver=0.21.0
-pkgrel=2
+pkgver=0.21.1
+pkgrel=1
 pkgdesc='Public Suffix List library (mingw-w64)'
 url='https://github.com/rockdaboot/libpsl'
 arch=(any)
 license=('MIT')
 depends=('mingw-w64-libidn2' 'mingw-w64-libunistring')
 makedepends=('mingw-w64-configure' 'python')
-source=(https://github.com/rockdaboot/libpsl/releases/download/${_pkgname}-${pkgver}/${_pkgname}-${pkgver}.tar.gz)
-sha512sums=('165c4f0b0640a813d512bd916e1532e32e43c8c81a5efd048f3a5b07b1b3c9129b4c4b5008b8b11a7c1b3914caea17564321389cd350bf1d687d53a97f2afa4d')
+source=(https://github.com/rockdaboot/libpsl/releases/download/${pkgver}/${_pkgname}-${pkgver}.tar.gz)
+sha512sums=('a5084b9df4ff2a0b1f5074b20972efe0da846473396d27b57967c7f6aa190ab3c910b4bfc4f8f03802f08decbbad5820d850c36ad59610262ae37fe77de0c7f5')
 
 _architectures="i686-w64-mingw32 x86_64-w64-mingw32"
 
@@ -39,9 +39,12 @@ package() {
   for _arch in ${_architectures}; do
     cd "${srcdir}/${_pkgname}-${pkgver}/build-${_arch}"
     make DESTDIR="${pkgdir}" install
-    find "$pkgdir/usr/${_arch}" -name '*.exe' -delete
+    find "$pkgdir/usr/${_arch}" -name '*.exe' -exec ${_arch}-strip --strip-all {} \;
     find "$pkgdir/usr/${_arch}" -name '*.dll' -exec ${_arch}-strip --strip-unneeded {} \;
-    find "$pkgdir/usr/${_arch}" -name '*.a' -o -name '*.dll' | xargs ${_arch}-strip -g
+    find "$pkgdir/usr/${_arch}" -name '*.a' -o -name '*.dll' -exec ${_arch}-strip -g {} \;
+    if [[ $NO_EXECUTABLES ]]; then
+      find "$pkgdir/usr/${_arch}" -name '*.exe' -delete
+    fi
   done
 }
